Fix lantern
