+++ src/features/island/collectibles/components/RadianceLantern.tsx
@@ -1,18 +1,18 @@
 import React from "react";
 
-import oceanLantern from "src/assets/sfts/ocean_lantern.png";
+import radianceLantern from "src/assets/sfts/radiance_lantern.webp";
 import { PIXEL_SCALE } from "features/game/lib/constants";
 
-export const OceanLantern: React.FC = () => {
+export const RadianceLantern: React.FC = () => {
   return (
     <div className="flex justify-center items-center pointer-events-none">
       <img
-        src={oceanLantern}
+        src={radianceLantern}
         style={{
           width: `${PIXEL_SCALE * 11}px`,
         }}
         className="paper-floating"
-        alt="Ocean Lantern"
+        alt="Radiance Lantern"
       />
     </div>
   );
